Atualizando fluxo das activitys para implementar CRUD

+++ app/src/main/java/br/com/matheush/appsqlite/dao/IDao.java
@@ -16,7 +16,7 @@ public interface IDao<T> {
     public void detetaTodos();
     public void altera(T objeto);
     public List<T> getObejetos();
-    public T getObejeto(int id);
+    public T getObejeto(long id);
